+++ src/main/java/net/frozenblock/wilderwild/client/WWBlockRenderLayers.java
@@ -67,7 +67,7 @@ public final class WWBlockRenderLayers {
 		renderLayerRegistry.putBlock(WWBlocks.FROZEN_BUSH, RenderType.cutout());
 		renderLayerRegistry.putBlock(WWBlocks.POTTED_FROZEN_BUSH, RenderType.cutout());
 
-		renderLayerRegistry.putBlock(WWBlocks.RED_CREEPING_MOSS, RenderType.cutout());
+		renderLayerRegistry.putBlock(WWBlocks.AUBURN_CREEPING_MOSS, RenderType.cutout());
 
 		renderLayerRegistry.putBlock(WWBlocks.DATURA, RenderType.cutout());
 		renderLayerRegistry.putBlock(WWBlocks.CATTAIL, RenderType.cutout());