+++ src/main/java/net/frozenblock/wilderwild/config/WWWorldgenConfig.java
@@ -183,8 +183,8 @@ public final class WWWorldgenConfig {
 		@EntrySyncData("mossDecoration")
 		public boolean mossDecoration = true;
 
-		@EntrySyncData("redMossDecoration")
-		public boolean redMossDecoration = true;
+		@EntrySyncData("auburnMoss")
+		public boolean auburnMoss = true;
 
 		@EntrySyncData("paleMossDecoration")
 		public boolean paleMossDecoration = true;
@@ -329,8 +329,8 @@ public final class WWWorldgenConfig {
 		@EntrySyncData("oceanMossGeneration")
 		public boolean oceanMossGeneration = true;
 
-		@EntrySyncData("oceanRedMossGeneration")
-		public boolean oceanRedMossGeneration = true;
+		@EntrySyncData("oceanAuburnMossGeneration")
+		public boolean oceanAuburnMossGeneration = true;
 	}
 
 	public static class BiomePlacement {