+++ src/main/java/net/frozenblock/wilderwild/config/gui/WWWorldgenConfigGui.java
@@ -862,15 +862,15 @@ public final class WWWorldgenConfigGui {
 			"mossDecoration",
 			configInstance
 		);
-		var redMossDecoration = FrozenClothConfig.syncedEntry(
-			entryBuilder.startBooleanToggle(text("red_moss_decoration"), modifiedSurfaceDecoration.redMossDecoration)
-				.setDefaultValue(defaultSurfaceDecoration.redMossDecoration)
-				.setSaveConsumer(newValue -> surfaceDecoration.redMossDecoration = newValue)
-				.setTooltip(tooltip("red_moss_decoration"))
+		var auburnMoss = FrozenClothConfig.syncedEntry(
+			entryBuilder.startBooleanToggle(text("auburn_moss_generation"), modifiedSurfaceDecoration.auburnMoss)
+				.setDefaultValue(defaultSurfaceDecoration.auburnMoss)
+				.setSaveConsumer(newValue -> surfaceDecoration.auburnMoss = newValue)
+				.setTooltip(tooltip("auburn_moss_generation"))
 				.requireRestart()
 				.build(),
 			surfaceDecorationClazz,
-			"redMossDecoration",
+			"auburnMoss",
 			configInstance
 		);
 		var paleMossDecoration = FrozenClothConfig.syncedEntry(
@@ -1009,7 +1009,7 @@ public final class WWWorldgenConfigGui {
 		var surfaceDecorationCategory = FrozenClothConfig.createSubCategory(entryBuilder, category, text("surface_decoration"),
 			false,
 			tooltip("surface_decoration"),
-			coarseDecoration, gravelDecoration, mudDecoration, packedMudDecoration, stoneDecoration, mossDecoration, redMossDecoration, paleMossDecoration,
+			coarseDecoration, gravelDecoration, mudDecoration, packedMudDecoration, stoneDecoration, mossDecoration, auburnMoss, paleMossDecoration,
 			scorchedSandDecoration, scorchedRedSandDecoration, sandstoneDecoration, clayDecoration, clearingDecoration, taigaBoulders,
 			snowPiles, fragileIceDecoration, icicleDecoration, lakes, basins
 		);
@@ -1191,22 +1191,22 @@ public final class WWWorldgenConfigGui {
 			"oceanMossGeneration",
 			configInstance
 		);
-		var oceanRedMossGeneration = FrozenClothConfig.syncedEntry(
-			entryBuilder.startBooleanToggle(text("ocean_red_moss_generation"), modifiedAquatic.oceanRedMossGeneration)
-				.setDefaultValue(defaultAquatic.oceanRedMossGeneration)
-				.setSaveConsumer(newValue -> aquatic.oceanRedMossGeneration = newValue)
-				.setTooltip(tooltip("ocean_red_moss_generation"))
+		var oceanAuburnMossGeneration = FrozenClothConfig.syncedEntry(
+			entryBuilder.startBooleanToggle(text("ocean_auburn_moss_generation"), modifiedAquatic.oceanAuburnMossGeneration)
+				.setDefaultValue(defaultAquatic.oceanAuburnMossGeneration)
+				.setSaveConsumer(newValue -> aquatic.oceanAuburnMossGeneration = newValue)
+				.setTooltip(tooltip("ocean_auburn_moss_generation"))
 				.requireRestart()
 				.build(),
 			aquaticClazz,
-			"oceanRedMossGeneration",
+			"oceanAuburnMossGeneration",
 			configInstance
 		);
 
 		var aquaticGenerationCategory = FrozenClothConfig.createSubCategory(entryBuilder, category, text("aquatic_generation"),
 			false,
 			tooltip("aquatic_generation"),
-			riverPool, algae, plankton, seagrass, spongeBud, barnacle, cattail, seaAnemone, seaWhip, tubeWorm, hydrothermalVent, oceanMossGeneration, oceanRedMossGeneration
+			riverPool, algae, plankton, seagrass, spongeBud, barnacle, cattail, seaAnemone, seaWhip, tubeWorm, hydrothermalVent, oceanMossGeneration, oceanAuburnMossGeneration
 		);
 
 		var transition = config.transitionGeneration;