+++ src/main/java/net/frozenblock/wilderwild/datagen/recipe/WWNaturalRecipeProvider.java
@@ -83,7 +83,7 @@ public final class WWNaturalRecipeProvider {
 			.unlockedBy(RecipeProvider.getHasName(WWBlocks.CATTAIL), provider.has(WWBlocks.CATTAIL))
 			.save(exporter, WWConstants.string(RecipeProvider.getConversionRecipeName(Items.STRING, WWBlocks.CATTAIL)));
 
-		provider.carpet(WWBlocks.RED_MOSS_CARPET, WWBlocks.RED_MOSS_BLOCK);
+		provider.carpet(WWBlocks.AUBURN_MOSS_CARPET, WWBlocks.AUBURN_MOSS_BLOCK);
 	}
 
 	private static void oneToOneConversionRecipe(RecipeProvider provider, RecipeOutput recipeOutput, ItemLike result, ItemLike ingredient, @Nullable String group) {