+++ src/main/java/net/frozenblock/wilderwild/datagen/tag/WWBlockTagProvider.java
@@ -467,7 +467,7 @@ public final class WWBlockTagProvider extends FabricTagProvider.BlockTagProvider
 			.add(Blocks.MOSS_BLOCK)
 			.add(Blocks.PACKED_MUD)
 			.add(Blocks.SNOW)
-			.add(WWBlocks.RED_MOSS_BLOCK);
+			.add(WWBlocks.AUBURN_MOSS_BLOCK);
 
 		this.getOrCreateTagBuilder(WWBlockTags.PACKED_MUD_REPLACEABLE)
 			.add(Blocks.STONE)
@@ -481,7 +481,7 @@ public final class WWBlockTagProvider extends FabricTagProvider.BlockTagProvider
 			.add(Blocks.SPONGE)
 			.add(Blocks.CLAY)
 			.add(Blocks.MOSS_BLOCK)
-			.add(WWBlocks.RED_MOSS_BLOCK)
+			.add(WWBlocks.AUBURN_MOSS_BLOCK)
 			.addOptionalTag(WWBlockTags.MESOGLEA);
 
 		this.getOrCreateTagBuilder(WWBlockTags.BASIN_REPLACEABLE)
@@ -489,7 +489,7 @@ public final class WWBlockTagProvider extends FabricTagProvider.BlockTagProvider
 			.add(Blocks.COARSE_DIRT)
 			.add(Blocks.PODZOL)
 			.add(Blocks.MOSS_BLOCK)
-			.add(WWBlocks.RED_MOSS_BLOCK);
+			.add(WWBlocks.AUBURN_MOSS_BLOCK);
 
 		this.getOrCreateTagBuilder(WWBlockTags.HYDROTHERMAL_VENT_REPLACEABLE)
 			.add(Blocks.CLAY)
@@ -957,9 +957,9 @@ public final class WWBlockTagProvider extends FabricTagProvider.BlockTagProvider
 			.add(WWBlocks.BARNACLES)
 			.add(WWBlocks.PHLOX)
 			.add(WWBlocks.LANTANAS)
-			.add(WWBlocks.RED_MOSS_BLOCK)
-			.add(WWBlocks.RED_MOSS_CARPET)
-			.add(WWBlocks.RED_CREEPING_MOSS);
+			.add(WWBlocks.AUBURN_MOSS_BLOCK)
+			.add(WWBlocks.AUBURN_MOSS_CARPET)
+			.add(WWBlocks.AUBURN_CREEPING_MOSS);
 
 		this.getOrCreateTagBuilder(BlockTags.MINEABLE_WITH_PICKAXE)
 			.add(WWBlocks.STONE_CHEST)
@@ -1034,8 +1034,8 @@ public final class WWBlockTagProvider extends FabricTagProvider.BlockTagProvider
 			.add(WWBlocks.FROZEN_FERN)
 			.add(WWBlocks.FROZEN_LARGE_FERN)
 			.add(WWBlocks.FROZEN_BUSH)
-			.add(WWBlocks.RED_MOSS_CARPET)
-			.add(WWBlocks.RED_CREEPING_MOSS)
+			.add(WWBlocks.AUBURN_MOSS_CARPET)
+			.add(WWBlocks.AUBURN_CREEPING_MOSS)
 			.addOptionalTag(WWBlockTags.NEMATOCYSTS);
 
 		this.getOrCreateTagBuilder(BlockTags.ENDERMAN_HOLDABLE)
@@ -1046,10 +1046,10 @@ public final class WWBlockTagProvider extends FabricTagProvider.BlockTagProvider
 			.add(WWBlocks.SCORCHED_RED_SAND);
 
 		this.getOrCreateTagBuilder(BlockTags.DIRT)
-			.add(WWBlocks.RED_MOSS_BLOCK);
+			.add(WWBlocks.AUBURN_MOSS_BLOCK);
 
 		this.getOrCreateTagBuilder(BlockTags.SNIFFER_DIGGABLE_BLOCK)
-			.add(WWBlocks.RED_MOSS_BLOCK);
+			.add(WWBlocks.AUBURN_MOSS_BLOCK);
 
 		this.getOrCreateTagBuilder(BlockTags.NEEDS_STONE_TOOL)
 			.add(WWBlocks.GABBRO)
@@ -1176,8 +1176,8 @@ public final class WWBlockTagProvider extends FabricTagProvider.BlockTagProvider
 		this.getOrCreateTagBuilder(BlockTags.COMBINATION_STEP_SOUND_BLOCKS)
 			.add(WWBlocks.POLLEN)
 			.add(WWBlocks.BARNACLES)
-			.add(WWBlocks.RED_MOSS_CARPET)
-			.add(WWBlocks.RED_CREEPING_MOSS);
+			.add(WWBlocks.AUBURN_MOSS_CARPET)
+			.add(WWBlocks.AUBURN_CREEPING_MOSS);
 
 		this.getOrCreateTagBuilder(BlockTags.FLOWER_POTS)
 			.add(WWBlocks.POTTED_BAOBAB_NUT)
@@ -1261,8 +1261,8 @@ public final class WWBlockTagProvider extends FabricTagProvider.BlockTagProvider
 		this.getOrCreateTagBuilder(BlockTags.MANGROVE_ROOTS_CAN_GROW_THROUGH)
 			.add(WWBlocks.ALGAE)
 			.add(WWBlocks.PLANKTON)
-			.add(WWBlocks.RED_MOSS_CARPET)
-			.add(WWBlocks.RED_CREEPING_MOSS);
+			.add(WWBlocks.AUBURN_MOSS_CARPET)
+			.add(WWBlocks.AUBURN_CREEPING_MOSS);
 
 		this.getOrCreateTagBuilder(BlockTags.GUARDED_BY_PIGLINS)
 			.add(WWBlocks.STONE_CHEST);
@@ -1391,8 +1391,8 @@ public final class WWBlockTagProvider extends FabricTagProvider.BlockTagProvider
 			.add(Blocks.BROWN_MUSHROOM)
 			.add(WWBlocks.ALGAE)
 			.add(WWBlocks.PLANKTON)
-			.add(WWBlocks.RED_MOSS_CARPET)
-			.add(WWBlocks.RED_CREEPING_MOSS)
+			.add(WWBlocks.AUBURN_MOSS_CARPET)
+			.add(WWBlocks.AUBURN_CREEPING_MOSS)
 			.addOptionalTag(BlockTags.SMALL_FLOWERS);
 
 		this.getOrCreateTagBuilder(WWBlockTags.FALLEN_TREE_STUMP_PLACEABLE_ON)